+++ utils/file/ls/stat_time_darwin.go
@@ -8,13 +8,13 @@ import (
 )
 
 func statAtime(stat *syscall.Stat_t) time.Time {
-	return time.Unix(stat.Atimespec.Sec, stat.Atimespec.Nsec)
+	return time.Unix(int64(stat.Atimespec.Sec), int64(stat.Atimespec.Nsec))
 }
 
 func statCtime(stat *syscall.Stat_t) time.Time {
-	return time.Unix(stat.Ctimespec.Sec, stat.Ctimespec.Nsec)
+	return time.Unix(int64(stat.Ctimespec.Sec), int64(stat.Ctimespec.Nsec))
 }
 
 func statBirthtime(stat *syscall.Stat_t) (time.Time, bool) {
-	return time.Unix(stat.Birthtimespec.Sec, stat.Birthtimespec.Nsec), true
+	return time.Unix(int64(stat.Birthtimespec.Sec), int64(stat.Birthtimespec.Nsec)), true
 }
