+++ utils/file/ls/stat_time_linux.go
@@ -8,11 +8,11 @@ import (
 )
 
 func statAtime(stat *syscall.Stat_t) time.Time {
-	return time.Unix(stat.Atim.Sec, stat.Atim.Nsec)
+	return time.Unix(int64(stat.Atim.Sec), int64(stat.Atim.Nsec))
 }
 
 func statCtime(stat *syscall.Stat_t) time.Time {
-	return time.Unix(stat.Ctim.Sec, stat.Ctim.Nsec)
+	return time.Unix(int64(stat.Ctim.Sec), int64(stat.Ctim.Nsec))
 }
 
 func statBirthtime(stat *syscall.Stat_t) (time.Time, bool) {
